backend/requirements.txt
fastapi
uvicorn[standard]
pinecone
google-genai
python-dotenv
pypdf